+++ apps/client/src/environments/game-env.ts
@@ -1,6 +1,6 @@
 export const gameEnv = {
   maxLevel: 80,
-  globalGameVersion: 5.401,
-  koreanGameVersion: 5.25,
+  globalGameVersion: 5.41,
+  koreanGameVersion: 5.3,
   chineseGameVersion: 5.301
 };